+++ R/has.invalid.multibyte.string.R
@@ -1,4 +1,4 @@
-#' Determine if "invalid multibyte string" error will be triggered due to wrong encoding.
+#' Internal function - Determine if "invalid multibyte string" error will be triggered due to wrong encoding.
 #'
 #' @param x String vector to be tested
 #' @param return.elements Logical value. If `TRUE`, then the function returns a logical value for each element of x, otherwise a single logical is returned for the entire vector.